Mongoid eviction notice
